% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_file_metadata.R
\name{bulk_update_file_metadata}
\alias{bulk_update_file_metadata}
\title{Updates a list of files based on their id and corresponding vectors}
\usage{
bulk_update_file_metadata(
  files,
  label = NULL,
  directoryLabel = NULL,
  description = NULL,
  restricted = NULL,
  categories = NULL,
  key = Sys.getenv("DATAVERSE_KEY"),
  server = "data.qdr.syr.edu"
)
}
\arguments{
\item{files}{A numerical vector. Dataverse IDs}

\item{label}{A character vector. Filenames for the files}

\item{directoryLabel}{A character vector. Folders for the files}

\item{description}{A character vector. Descriptions for files}

\item{restricted}{A boolean vector. Whether files are restricted}

\item{categories}{A character vector. File tags. Mulitple tags are comma or semicomma separated}

\item{key}{The API key for the dataverse installation}

\item{server}{The dataverse installation}
}
\description{
The assumption here is that this is based on a dataframe/spreadsheet so that file IDs and metadata vectors are guaranteed to be aligned
}
\examples{
\dontrun{
bulk_update_file_metadata(
  c(402881, 395305),
  description = c("new description", "other new description"),
  categories = c("Data", "Documentation"),
  server = "demo.dataverse.org"
)
}

}
\seealso{
update_file_metadata
}
