% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folder}
\alias{create_folder}
\title{Creates a download directory named QDR Project - Author with two subfolders in the specified folder}
\usage{
create_folder(data, folder = getwd())
}
\arguments{
\item{data}{A dataverse dataset object}

\item{folder}{The parent folder in which the download directory will be created. Typically the Dropbox folder.}
}
\value{
The filepath of the created folder
}
\description{
Creates a download directory named QDR Project - Author with two subfolders in the specified folder
}
\examples{
\dontrun{
data <- get_dataset("doi:10.5064/F6NUVQRR")
folderpath <- create_folder(data, folder="C:/Users/Sebastian/Dropbox/QDR Project - Smith")
}
}
