% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_byDate.R
\name{datasets_byDate}
\alias{datasets_byDate}
\title{Produces an html file with all datasets published in a given period}
\usage{
datasets_byDate(
  startDate = "2016-01-01",
  endDate = Sys.Date(),
  citationsDir = tempdir(),
  pageName = "citations.html"
)
}
\arguments{
\item{startDate}{The (inclusive) start date in YYYY-MM-DD format. default 2016-01-01}

\item{endDate}{The (exclusive) end date in YYYY-MM-DD format. default: today's date using `Sys.Date()`}

\item{citationsDir}{directory for the new file to be created. Defaults to `tempdir()`}

\item{pageName}{name of the html file. defaults to citations.html}
}
\value{
None. Writes an html file to local disk
}
\description{
Produces an html file with all datasets published in a given period
}
\examples{
\dontrun{
datasets_byDate("2017-01-01", "2017-06-01")
}
}
