% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{download_files}
\alias{download_files}
\title{Returns the zip file of all files in a dataset as a raw vector}
\usage{
download_files(data)
}
\arguments{
\item{data}{A dataverse dataset object}
}
\value{
The .zip file of all datafiles as a raw vector
}
\description{
Returns the zip file of all files in a dataset as a raw vector
}
\examples{
\dontrun{
data <- get_dataset("doi:10.5064/F6NUVQRR")
zip <- download_files(data)
}
}
