% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_filelist.R
\name{export_filelist}
\alias{export_filelist}
\title{Writes a CSV file of all files pertaining to the current DOI to the specified folder}
\usage{
export_filelist(data, folder = getwd())
}
\arguments{
\item{doi}{A string. The DOI of a dataset in the form doi:10.1234/5678}
}
\value{

}
\description{
Writes a CSV file of all files pertaining to the current DOI to the specified folder
}
\examples{
\dontrun{
export_files("doi:10.5064/F6NUVQRR")
}
}
