% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_author}
\alias{get_author}
\title{Extracts the first author last name}
\usage{
get_author(data)
}
\arguments{
\item{data}{a dataverse API object for a dataset}
}
\value{
author the last name of the author
}
\description{
Extracts the first author last name
}
\examples{
\dontrun{
get_author(data)
}
}
