% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_email}
\alias{get_email}
\title{Returns the email of the first listed contact of the}
\usage{
get_email(data)
}
\arguments{
\item{data}{a dataverse API object for a dataset}
}
\value{
email address of the first listed contact
}
\description{
Returns the email of the first listed contact of the
}
\examples{
\dontrun{
get_email(data)
}
}
