% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_file_metadata.R
\name{get_file_json}
\alias{get_file_json}
\title{Retrieves file metadata information for a file given its ID}
\usage{
get_file_json(
  file,
  key = Sys.getenv("DATAVERSE_KEY"),
  server = "data.qdr.syr.edu"
)
}
\arguments{
\item{file}{The numeric ID for the file}

\item{key}{The API key}

\item{server}{The dataverse installation to be used}
}
\value{
The JSON return from the API as a List
}
\description{
Retrieves file metadata information for a file given its ID
}
\examples{
fileJSON <- get_file_json(396501, key="", server = "demo.dataverse.org")
}
