% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_filelist.R
\name{get_filelist}
\alias{get_filelist}
\title{Extracts a list of filenames with selected characteristics from a dataverse dataset object}
\usage{
get_filelist(data)
}
\arguments{
\item{data}{A list. The response of the dataverse API to a request for dataset metadata}
}
\value{
A data frame containing all filenames with selected characteristics
}
\description{
Extracts a list of filenames with selected characteristics from a dataverse dataset object
}
\examples{
\dontrun{
get_filelist(data)
}
}
