% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_shortTitle}
\alias{get_shortTitle}
\title{Returns a shortened title}
\usage{
get_shortTitle(title)
}
\arguments{
\item{title}{A string, typically a dataset title.}
}
\value{
short_title, a string.
}
\description{
Removes "Data for:" and "Replication Data for:" and uses the first five words for the title or the title up to the first colon, whichvever comes first.
}
\examples{
short_title <- get_shortTitle(
  "Data for: From Pews to Politics: Religious Sermons and Political Participation in Africa"
  )

}
