% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_title}
\alias{get_title}
\title{Extracts the title of the datasets from its object}
\usage{
get_title(data)
}
\arguments{
\item{data}{}
}
\value{
The title of the dataset. A string
}
\description{
Extracts the title of the datasets from its object
}
\examples{
\dontrun{
get_title(data)
}
}
