% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_url}
\alias{get_url}
\title{Creates the QDR link for a dataset}
\usage{
get_url(doi)
}
\arguments{
\item{doi}{A string. The doi of a davaser dataset in the form doi:10.1234/5678}
}
\value{
the QDR url for a dataset. A string.
}
\description{
Creates the QDR link for a dataset
}
\examples{
url <- get_url("doi:10.5064/F6NUVQRR")
}
