% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_curate.R
\name{gh_curate}
\alias{gh_curate}
\title{Creates issues and projects in the QDR Github repository}
\usage{
gh_curate(doi, owner = "QualitativeDataRepository", repo = "Project-Curation")
}
\arguments{
\item{doi}{}
}
\value{

}
\description{
Creates issues and projects in the QDR Github repository
}
\examples{
\dontrun{
gh_curate("doi:10.5064/F6NUVQRR")
}
}
