% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{save_files}
\alias{save_files}
\title{Save a downloaded zip file to the appropriate folder}
\usage{
save_files(zip, data, folder = getwd())
}
\arguments{
\item{zip}{A downloaded zip file as a raw vector}

\item{data}{A dataverse dataset object}

\item{folder}{A folder, typically the project's folder on Dropbox}
}
\description{
Saves the zip file in the "/Original Deposit" folder in the project folder, and extracts the zip file, minus the Manifest, into the "/QDR Prepared" subfolder
}
\examples{
\dontrun{
data <- get_dataset("doi:10.5064/F6NUVQRR")
zip <- download_files(data)
save_files(zip, data)
}
}
