% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_curation.R
\name{start_curation}
\alias{start_curation}
\title{Initiaties a curation project: Creates gh projects, DB folder, downloads files & filelist}
\usage{
start_curation(
  doi,
  folder = getwd(),
  owner = "QualitativeDataRepository",
  repo = "Project-Curation"
)
}
\arguments{
\item{doi}{A string. the dataset DOI in the form "doi:10.1234/5667"}

\item{folder}{A path to a folder, typically the local Dropobox folder}

\item{owner}{A string. The owner of the github repository to be used}

\item{repo}{A string. The name of the github repository to be used}
}
\description{
Initiaties a curation project: Creates gh projects, DB folder, downloads files & filelist
}
\examples{
\dontrun{
start_curation("doi:10.5064/F6NUVQRR", folder = "C:/Users/Sebastian/Dropbox (QDR)")
}
}
