% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_file_metadata.R
\name{update_file_metadata}
\alias{update_file_metadata}
\title{Updates the file metadata of a given file}
\usage{
update_file_metadata(
  file,
  label = NULL,
  directoryLabel = NULL,
  description = NULL,
  restricted = NULL,
  categories = NULL,
  key = Sys.getenv("DATAVERSE_KEY"),
  server = "data.qdr.syr.edu"
)
}
\arguments{
\item{file}{The numeric ID of a file}

\item{label}{A string. filename/label of the file.}

\item{directoryLabel}{A string. The directory for the file _not_ starting with a slash}

\item{description}{A string. The description of the file}

\item{restricted}{Boolean. TRUE (for restricted), FALSE for not.}

\item{categories}{A string. Tags for the file, comma or semicolon separated.}

\item{key}{The API key for the dataverse installation}

\item{server}{The dataverse installation to be used}
}
\value{
The API response (as text)
}
\description{
It will preserve all metadata that's not specifically addressed
}
\examples{
\dontrun{
# Updates the file description
update_message <- update_file_metadata(396501, description = "New file description")
}
}
