% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_filelist.R
\name{export_filelist}
\alias{export_filelist}
\title{Writes a CSV file of all files pertaining to the current DOI to the specified folder}
\usage{
export_filelist(data, folder = getwd())
}
\arguments{
\item{data}{A list. The output of dataverse::get_dataset()}

\item{folder}{A filepath. The folder for the file to be saved}
}
\description{
Writes a CSV file of all files pertaining to the current DOI to the specified folder
}
\examples{
\dontrun{
export_files(get_dataset("doi:10.5064/F6NUVQRR"))
}
}
