% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_url}
\alias{get_url}
\title{Creates the QDR link for a dataset}
\usage{
get_url(doi)
}
\arguments{
\item{doi}{}
}
\value{
the QDR url for a dataset. A string.
}
\description{
Creates the QDR link for a dataset
}
\examples{
url <- get_url("doi:10.5064/F6NUVQRR")
}
