% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_curate.R
\name{gh_curate}
\alias{gh_curate}
\title{Creates issues and projects in the QDR Github repository}
\usage{
gh_curate(doi, owner = "QualitativeDataRepository", repo = "Project-Curation")
}
\arguments{
\item{doi}{A string. The doi of a davaser dataset in the form doi:10.1234/5678}

\item{owner}{A string. The owner of the github repository}

\item{repo}{A string. The name of the github repo}
}
\description{
Creates issues and projects in the QDR Github repository
}
\examples{
\dontrun{
gh_curate("doi:10.5064/F6NUVQRR")
}
}
